+++ coteries/croisadefranque.h
@@ -11,7 +11,7 @@ public:
     void GenererTraitCompatibles();
     QString GetNom();
     QString GetId();
-    float GetPoidsDemo();
+    double GetPoidsDemo();
 
     QString GetDescription();
     QString GetDescriptionUniversite();
